requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
djangorestframework==3.15.1
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
priority==1.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
redis==5.0.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
zope.interface==6.3
